* 'ff' of 141.65.6.106:ogs_sources:
  fixed some compiler warnings
  added missing files
  fixed bug concerning conversion of vtk-meshes to ogs-meshes

+++ DataView/DiagramView/DiagramScene.cpp
@@ -167,7 +167,7 @@ void DiagramScene::constructGrid()
 	_grid = addGrid(_bounds, numXTicks, numYTicks, pen);
 
 	for (int i = 0; i <= numXTicks; ++i) {
-		float x = _bounds.left()/_scaleX + (i * (_bounds.width()/_scaleX) / numXTicks);
+		int x = static_cast<int>(_bounds.left()/_scaleX + (i * (_bounds.width()/_scaleX) / numXTicks));
 		QDateTime currentDate = _startDate.addDays(x);
 		_xTicksText.push_back(addNonScalableText(currentDate.toString("dd.MM.yyyy")));
 		_xTicksText.last()->setPos(x*_scaleX, _bounds.bottom() + 15);
